requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
nltk==3.8.1
python-multipart==0.0.6
google-generativeai==0.3.2